+++ ShandalarImageDecoder.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.329
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29201.188
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PicDecode", "PicDecode\PicDecode.csproj", "{5ADEBD2B-0391-44AB-9484-63A427C957D6}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ShandalarImageDecoder", "ShandalarImageDecoder\ShandalarImageDecoder.csproj", "{5ADEBD2B-0391-44AB-9484-63A427C957D6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
