+++ ShandalarImageDecoder/App.config
