+++ ShandalarImageDecoder/Form1.cs
@@ -13,7 +13,7 @@ using System.Windows.Forms;
 using System.IO;
 using System.Runtime.InteropServices;
 
-namespace PicDecode
+namespace ShandalarImageDecoder
 {
     public partial class Form1 : Form
     {
