+++ ShandalarImageDecoder/Form1.resx
@@ -132,4 +132,72 @@
   <metadata name="openFileDialog3.TrayLocation" type="System.Drawing.Point, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a">
     <value>601, 17</value>
   </metadata>
+  <assembly alias="System.Drawing" name="System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a" />
+  <data name="$this.Icon" type="System.Drawing.Icon, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
+    <value>
+        AAABAAEAMDAAAAAAAACoDgAAFgAAACgAAAAwAAAAYAAAAAEACAAAAAAAAAkAAAAAAAAAAAAAAAEAAAAA
+        AAD///8AQkJCADExMQBKSlIAQkJKADk5QgAxMTkAKSkxAEpKWgAhISkAOTlKABgYIQAQEBgAGBgpACEh
+        OQAICBAAISFSABAQKQAhIVoAGBhCABAQMQAYGFIAEBA5ABAQQgAICCEAGBhrABAQSgAYGHMAEBBSABAQ
+        YwAICDEACAhCAAgISgAICFIACAhjAAAAcwAAAGMAAABaAAAAOQAAADEAAAApAAAAIQAAAAgAISmEACEp
+        awAhISEAGCFaAHN7lABrc4wAY2uEAAgQKQBze4wAa3OEAFJaawBKUmMAjJy9AISUtQAIECEAY2t7AJyt
+        zgCUpcYAQkpaADk5OQBaY3MAOUJSAIyctQCElK0Aa3uUAGNzjACcrcYAlKW9AHuMpQBzhJwAnLXWAJSt
+        zgAYGBgAUlpjAISUpQBre4wAY3OEADE5QgCEnLUASlprABAYIQBrc3sASlJaAEJKUgB7jJwAc4SUAJy1
+        zgCUrcYAjKW9AFprewApMTkAe5StAFJjcwBzjKUAISkxAFpjawA5QkoAGCEpAIScrQB7lKUAjKW1AHOM
+        nABrhJQAa3uEAEpaYwBjc3sAQlJaAHuMlABaa3MAUmNrADlKUgAxQkoAa3NzAGNrawBSWloASlJSAEJK
+        SgB7jIwAc4SEAGt7ewA5QkIAY3NzADE5OQBaa2sAKTExAFJjYwAhKSkAGCEhABAYGAAIEBAAAAgIAFpr
+        YwAQEBAAa3tzAGNzawBKWlIAQlJKACkpKQBSWlIAWmNaAEpSSgBCSkIAOUI5ADE5MQApMSkAUmNSAEJS
+        QgAhKSEAOUo5AEJSOQAICAgASlJCADE5KQApMSEAEBgIAAgQAABCSjkAGCEIABAYAAApMRgAISkQAEpS
+        MQBCQjkAOTkxADExKQAhIRgAOTkpADExIQAYGBAAKSkYABAQCAAhIRAAKSkQABgYCAAhIQgAKSkIABgY
+        AAAQEAAACAgAAAgAQgAQCFIACABrAAgA/wAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
+        AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
+        AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
+        AAAAAAAAAAAAAAAAAAB1Aa89bodWA7cErj0qAFmBJIgAAO8FMAAAAAQAKIj3GdcFehrXBRgAAAAAADAA
+        MAAwAAEIAADvSigAAAAwAAAAMAAAAAEACAAAAGBAYwD4IAAAsodQfxcBU3/4ICl5Uriq9ar17wX0h6wR
+        nwSyDfYFurq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6
+        urq6urq6urq6urigobq6urq6urq6urq6uqG6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urF2
+        sLq6urNTurq6urq6upSqurq6urq6urq6urq6urq6urq6urq6urq6urq6urq6HJEEq7q6qZq6urq6urq6
+        LHp7urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6tGN3sLoNjZS6uiaklbq6r0Omurq6urq6urq6
+        urq6urq6urq6urq6urq6urq6urq6mQV3SydjNY66uiF8ZT8ofU6Surq6urq6urq6urq6urq6urq6urq6
+        urq6urq6urq6tHt3S0tVP5S6uhmIQjq6Y0OSurq6urq6urq6urq6urq6urq6urq6urq6urq6urq6tHsE
+        gVVwPz+buit5R2yEVk5/urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6tHt3Azo6P282gVVYWC81
+        XDCcurq6urq6urq6urq6urq6urq6urq6urq6urq6urq6mQUEVWJPb2I/WEdOTmhCaE+curq6urq6urq6
+        urq6urq6urq6urq6urq6urq6urq6tAV7BFU/OnA/WFgwTlhDMEyMuLq6urq6urq6urq6urq6urq6urq6
+        urq6urq6urocAnt7VUxVTD9iTkNYV0hDb1YFoCK6urq6urq6urq6urq6urq6urq6uronHxoVEhKnewUE
+        cD99g1Y1bE5OOmhIXFVWBIwSEhIVFx+6urq6urq6urq6urq6F12nAqY+kaVjewRVVVYFug9MQjCDgTBO
+        T281a1WPj3aLkJEOILq6urq6urq6uroTpnt7e3t7BHdWVlVVVlZ3g7qDX3+6UE86XD9vP3A1azZVNlUD
+        myG6urq6urq6uhObewUFewUFewRWVnt7BFZVB7q6urq6a05cNVVVTDZ2VlZWVVVtA6K6urq6urq6uq8E
+        BXs+An19fQU+fQRWVgNVBbq6urqZT05cbzVWBX0GAn0FY3cIVpAVurq6urq6Fj57BX1WTD9PM04/cDE0
+        T281Y7WZg7oLQ1hOME40b09qTjQ6TFVWVVWcHrq6urpTp3s+BVVMYjZWTE9NZVdYMzQ/d5kMVgxdVy9O
+        WGhnO0JPXDVxbVxcNVVWqB66urqqBH1WcIEPD7q6ug+BP1dNWE58VoeDTEtQR2lIQkVCNoOZmbq6mQtM
+        OjVrBLC6uocFez4DCbq6urq6urq6ugw9TlhsTIcMcAtxR01bWU8Murq6urq6urqFCUxVBJO6ulORBWOB
+        urq6urq6urq6urq6UDpPYi2DYl1cUUZbNYW6urq6urq6urq6mWM2cJu6urqqewR/Krq6urq6urq6urq6
+        jARVNWMFak9YQkhcBrq6urq6urq6urq6umNcYhW6urqgewR9urq6urq6urq6urqCBZEFBGJDZkZmOlVj
+        BUu6urq6urq6urq6mW1Ojbq6urqxewQFmbq6urq6urq6mYI+AVA+fTZNPDxPVgUFez6Cmbq6urq6urq6
+        hD9OgB26urqoAXsEgbq6urq6hYctAgV7UAQIVmN8ZjR7PndVVVVWBoe6urq6urq6XVhEb6G6urqie3tj
+        dwu6upkLAgUEVndjAzVvXGI1OlZ9YzVvb1xOTjUHhLq6urqZbEhPOku6urqvBAUEBAMDVTZMVVVwP3Bi
+        XE80TlhYTk9PNE4vTi9HQWdXTzZ7Y2sxYENEXIK6urqoe3t7d1VPTjQ6XDo6OjpPTi9YV0dmTUJCR1dH
+        ZmZNUVs8WlpaWlpRR0NOjrq6urqjAXsEBFY1fHBibzpsNE4zbldmQmVCQUJBQWVBQVFBQUFnN0E3UUJH
+        SGlOlLq6urqzY3tjd1Z1NUw1Pzo0TlhIR0JlZUFBQWdnW1tGW0ZbW0FRUWVCQl5XSFhDlbq6uroupnsE
+        BFZVVUxMcD86TlhXTUJBQUFBQTdGW0ZGRjxGW0FRQUJeQkdoSGlOqrq6urq6snd7d3dVVUxMYm9sTlhH
+        TUJBQWdGZ1tGRkZaWlpaRltBUVFCXkdoSFiGurq6urq6uqkEBHdVVVU1YjpPL1dNQmVnQUFnW0ZGWkVa
+        WkVaWjxbQUJCZkdoSFiYurq6urq6upx3YwRWVUxMPzpOL1dNZUFBZ1tGRkY8RlpaO0VZWjxbUTheR0dI
+        SIm6urq6urq6ui6pd3cIVUxiP2xObldCQUFGW0ZGRlpaWkVFWVk7RVpbN2VCZkdIV226urq6urq6urqM
+        kFZVTDViXDRYZkJlQUFbRjxaWkVaO0VaRllJWlpbQVFeR0dmiSy6urq6urq6uroUmwhVTDVvNE8xOm8/
+        SFFBQUZGWkVFRVlZQUdaSVtbUVFCZmZYm7q6urq6urq6urq6nZBVTD80VQuCYzZcTUFRQUFGRlpaWlo7
+        SldfL1pGUUJeXld3urq6urq6urq6urq6uqF3YmJQhIw1Q1dCQmVlZUFRQVtGW0ZbRlsvbUAxUVFlR4u6
+        urq6urq6urq6urq6urqddn26XWlYbkhXV0dmQkJCZVFRQVFRUVFbQm2ZXU5Yl7q6urq6urq6urq6urq6
+        urq6uoOFijNOTlhYaFdXR2ZNQkJeZUJlQmVCZ1GLumRLurq6urq6urq6urq6urq6urq6urq6uqSIiFhY
+        WG5XV0dHZkJNTWVNTWVleHmpuLq6urq6urq6urq6urq6urq6urq6urq6urq6uqSJiHmIeW5mZk14eHl5
+        eY6Yurq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6uqKpmKm6urq6urq6urq6urq6urq6
+        urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6
+        urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6
+        urq6urq6urq6urq6urq6urq6urq6urq6urq6urq6////////urr//x/+//+6uv//HP5//7q6//4Z/H//
+        urr//hGMf/+6uv/+AYB//7q6//4BgH//urr//gCAf/+6uv/+AAB//7q6//4AAH//uqH//gAAP/+6uv/8
+        AAAf/7q6/wAAAAB/urr8AAAAAB+6uvgAAAAAD7NT8AAAAAAPupTwAAAAAAe6uuAAAAAAA7q6wAAAAAAB
+        urrAAAAAAAG6HIAAAAAAAZq6gAAAAAABLHrAAAAAAAG6usAAAAAAA7q6wAAAAAABurrAAAAAAAG6tMAA
+        AAAAAZS6wAAAAAABr0PAAAAAAAO6usAAAAAAA7q6wAAAAAADurrAAAAAAAO6meAAAAAAB4668AAAAAAH
+        fU7wAAAAAA+6uvAAAAAAD7q6+AAAAAAPurr4AAAAAB+6tPwAAAAAP5S6/gAAAAB/Y0P/AAAAAP+6uv/A
+        AAAB/7q6//gAAAf/urr//wAAf/+6tP///D///z+b////////Vk7///////+6uv///////7q6
+</value>
+  </data>
 </root>
