+++ ShandalarImageDecoder/FormPalette.cs
@@ -8,7 +8,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace PicDecode
+namespace ShandalarImageDecoder
 {
     public partial class FormPalette : Form
     {
