+++ ShandalarImageDecoder/FormPalette.resx
