+++ ShandalarImageDecoder/PalDecoder.cs
@@ -9,7 +9,7 @@ using System.Threading.Tasks;
 using System.Drawing;
 using System.Text.RegularExpressions;
 
-namespace PicDecode
+namespace ShandalarImageDecoder
 {
     class PalDecoder
     {
