+++ ShandalarImageDecoder/Program.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace PicDecode
+namespace ShandalarImageDecoder
 {
     static class Program
     {
