+++ ShandalarImageDecoder/Properties/AssemblyInfo.cs
