+++ ShandalarImageDecoder/Properties/Resources.resx
