+++ ShandalarImageDecoder/Properties/Settings.settings
